refactor(core): decompose monolithic types.py into focused modules (#56)
### 1. GSoC Context & Goal 🎯

This PR improves codebase maintainability by decomposing the 1150-line monolithic `core/types.py` into 11 focused modules, establishing cleaner module boundaries as the library approaches its first stable API form in the command pipeline architecture.

### 2. The Change: What & Why 🛠️

Broke down the massive `core/types.py` into 11 specialized modules with clear separation of concerns: API parts, command states, conversation structures, result types, and validation utilities. The original `types.py` now serves as a compatibility shim that re-exports all types, ensuring full backward compatibility while dramatically improving code navigation and maintainability for the upcoming API stabilization phase.

### 3. How to Self-Verify 🧪

1. Run `make test-progressive` to verify all imports and functionality remain intact
2. Test backward compatibility with `python -c "from gemini_batch.core.types import Source, InitialCommand, ResultEnvelope; print('All imports work')"`
3. Verify the module structure with `python -c "import gemini_batch.core; print([m for m in dir(gemini_batch.core) if not m.startswith('_')])"`
4. Run `make lint` to confirm code organization meets quality standards
5. Execute existing examples to ensure no import breakage occurred

### 4. Impact & Next Steps 🚀

The codebase now has a clean, focused module structure that will significantly ease navigation during the final API hardening and documentation passes. This refactoring removes a major maintenance bottleneck and prepares the core types for stable public release while preserving all existing functionality.

### 5. Key Files Changed

- `src/gemini_batch/core/types.py`: Refactored into compatibility shim re-exporting all types
- `src/gemini_batch/core/api_parts.py`: Provider-agnostic API payload types
- `src/gemini_batch/core/commands.py`: Pipeline command state management
- `src/gemini_batch/core/result_envelope.py`: User-facing result structures
- `src/gemini_batch/core/sources.py`: Enhanced source handling with directory utilities
